+++ NEWS.md
@@ -1,8 +1,13 @@
 # NEWS
 
+# TODO:
+
+* Make sure values in interactive table are only modifed as allowed (e.g. values in range) 
 
 ## 0.2
-* modify grid
+
+* new table to modify grid ratings and constructs interactively
+* Bettin feature to modify grid, e.g. swap poles, remove constructs or elements
 
 ## 0.1